requirements.txt
fastapi~=0.111.0
uvicorn[standard]
mysql-connector-python~=8.4.0
pydantic~=2.7.4
bcrypt